完成运行bot代码的微服务

+++ backend/main-server/src/main/java/com/kob/mainserver/util/AuthenticationUtils.java
@@ -27,7 +27,7 @@ public class AuthenticationUtils {
         long userId;
         try {
             Claims claims = JwtUtils.parseJWT(token);
-            userId = new Integer(claims.getSubject());
+            userId = new Long(claims.getSubject());
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
